ayurveda-app/bot-brain/requirements.txt
PyPDF2>=3.0.0
pdfplumber
spacy
sentence-transformers
faiss-cpu

# Optional for advanced relation extraction / LLM
# transformers
# openai (if you use remote APIs)

# Backend API dependencies
fastapi>=0.95.2
uvicorn[standard]>=0.22.0
google-genai
motor>=3.3.0
pymongo>=4.6.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
